Tried and failed to get gradient diagrams to work, + refactoring
In my test data (recorded with Komoot), every GPX waypoint has an
elevation but the data only changes on every 10th waypoint or so, which
equals roughly every 100m (at least in my test data). This means that
the gradient calculations often result in exactly 0, because the
recorded elevation between data points is precisely equal.

This feature might still be useful for users with different GPS
recorders, but for working as I want it to, some major maths wizardry is
necessary and I will probably do more important stuff first.